Refactor agent module splits
