+++ src-tauri/src/agent/session_manager/compact.rs
@@ -1,23 +1,28 @@
-use super::AgentSessionManager;
 use crate::agent::compact_recovery::handle_compact_error_state;
 use crate::agent::events::AgentEventDispatcher;
 use crate::agent::llm::types::CompactStatePhase;
 use crate::agent::state::{AgentSession, CompactionResumeAction, DeferredWorkflowStep};
 use crate::agent::tauri_events::emit_compact_finished;
 use crate::repositories::{CompactContextRecord, SessionRepository, SessionStatus};
+use crate::repositories::compact_context_repository::CompactContextRepository;
+use crate::mcp::MCPServiceProxyManager;
+use tauri::AppHandle;
 use std::collections::HashMap;
 use std::sync::Arc;
 use tokio::sync::RwLock;
 
 fn spawn_resume_completion(
-    manager: &AgentSessionManager,
+    session_repo: &Arc<dyn SessionRepository>,
+    active_sessions: &Arc<RwLock<HashMap<String, AgentSession>>>,
+    proxy_manager: &Arc<MCPServiceProxyManager>,
+    app_handle: &AppHandle,
     session_id: &str,
     log_label: &'static str,
 ) {
-    let session_repo = manager.session_repo.clone();
-    let active_sessions = manager.active_sessions.clone();
-    let proxy_manager = manager.proxy_manager.clone();
-    let app_handle = manager.app_handle.clone();
+    let session_repo = session_repo.clone();
+    let active_sessions = active_sessions.clone();
+    let proxy_manager = proxy_manager.clone();
+    let app_handle = app_handle.clone();
     let resume_session_id = session_id.to_string();
 
     tokio::spawn(async move {
@@ -41,14 +46,17 @@ fn spawn_resume_completion(
 }
 
 fn spawn_resume_tool_execution(
-    manager: &AgentSessionManager,
+    session_repo: &Arc<dyn SessionRepository>,
+    active_sessions: &Arc<RwLock<HashMap<String, AgentSession>>>,
+    proxy_manager: &Arc<MCPServiceProxyManager>,
+    app_handle: &AppHandle,
     session_id: &str,
     tool_calls: Vec<crate::agent::types::ToolCall>,
 ) {
-    let session_repo = manager.session_repo.clone();
-    let active_sessions = manager.active_sessions.clone();
-    let proxy_manager = manager.proxy_manager.clone();
-    let app_handle = manager.app_handle.clone();
+    let session_repo = session_repo.clone();
+    let active_sessions = active_sessions.clone();
+    let proxy_manager = proxy_manager.clone();
+    let app_handle = app_handle.clone();
     let resume_session_id = session_id.to_string();
 
     tokio::spawn(async move {
@@ -65,14 +73,16 @@ fn spawn_resume_tool_execution(
 }
 
 async fn finalize_workflow_completion(
-    manager: &AgentSessionManager,
+    session_repo: &Arc<dyn SessionRepository>,
+    active_sessions: &Arc<RwLock<HashMap<String, AgentSession>>>,
+    app_handle: &AppHandle,
     session_id: &str,
     reason: crate::agent::events::WorkflowCompletionReason,
 ) -> Result<(), String> {
     crate::agent::lifecycle::update_session_status(
-        &manager.session_repo,
-        &manager.active_sessions,
-        &manager.app_handle,
+        session_repo,
+        active_sessions,
+        app_handle,
         session_id,
         SessionStatus::Idle,
     )
@@ -82,7 +92,7 @@ async fn finalize_workflow_completion(
         session_id: session_id.to_string(),
         reason,
     };
-    crate::agent::tauri_events::emit_agent_event(&manager.app_handle, event)
+    crate::agent::tauri_events::emit_agent_event(app_handle, event)
         .map_err(|e| format!("Failed to emit event: {}", e))
 }
 
@@ -96,15 +106,32 @@ pub async fn handle_compact_error_with_dispatcher(
     handle_compact_error_state(session_repo, active_sessions, dispatcher, session_id, error).await
 }
 
+pub struct CompactResponseParams<'a> {
+    pub active_sessions: &'a Arc<RwLock<HashMap<String, AgentSession>>>,
+    pub app_handle: &'a AppHandle,
+    pub session_repo: &'a Arc<dyn SessionRepository>,
+    pub proxy_manager: &'a Arc<MCPServiceProxyManager>,
+    pub session_id: &'a str,
+    pub from_id: String,
+    pub to_id: String,
+    pub summary: String,
+}
+
 pub async fn handle_compact_response(
-    manager: &AgentSessionManager,
-    session_id: &str,
-    from_id: String,
-    to_id: String,
-    summary: String,
+    params: CompactResponseParams<'_>,
 ) -> Result<(), String> {
+    let CompactResponseParams {
+        active_sessions,
+        app_handle,
+        session_repo,
+        proxy_manager,
+        session_id,
+        from_id,
+        to_id,
+        summary,
+    } = params;
     let (compaction, session_name) = {
-        let active = manager.active_sessions.read().await;
+        let active = active_sessions.read().await;
         active.get(session_id).map_or((None, None), |session| {
             (
                 Some(session.compaction.clone()),
@@ -138,10 +165,10 @@ pub async fn handle_compact_response(
         summary,
         created_at: chrono::Utc::now().timestamp_millis(),
     };
-    manager.save_compact_context(session_id, record).await?;
+    save_compact_context(active_sessions, session_id, record).await?;
 
     let resume_action = {
-        let active = manager.active_sessions.read().await;
+        let active = active_sessions.read().await;
         if let Some(session) = active.get(session_id) {
             Some(session.compaction.complete_success().await)
         } else {
@@ -160,7 +187,7 @@ pub async fn handle_compact_response(
         }
     );
     if let Err(error) = emit_compact_finished(
-        &manager.app_handle,
+        app_handle,
         session_id.to_string(),
         session_name,
         CompactStatePhase::Succeeded,
@@ -181,7 +208,7 @@ pub async fn handle_compact_response(
                     "▶️ Resuming deferred LLM completion after compaction for session {}",
                     session_id
                 );
-                spawn_resume_completion(manager, session_id, "deferred LLM completion");
+                spawn_resume_completion(session_repo, active_sessions, proxy_manager, app_handle, session_id, "deferred LLM completion");
             }
             DeferredWorkflowStep::ExecuteToolCalls {
                 assistant_message_id,
@@ -195,7 +222,7 @@ pub async fn handle_compact_response(
                 );
 
                 {
-                    let mut active = manager.active_sessions.write().await;
+                    let mut active = active_sessions.write().await;
                     if let Some(session) = active.get_mut(session_id) {
                         let expected_tool_call_ids: std::collections::HashSet<String> =
                             tool_calls.iter().map(|tc| tc.id.clone()).collect();
@@ -213,10 +240,10 @@ pub async fn handle_compact_response(
                             });
                     }
                 }
-                spawn_resume_tool_execution(manager, session_id, tool_calls);
+                spawn_resume_tool_execution(session_repo, active_sessions, proxy_manager, app_handle, session_id, tool_calls);
             }
             DeferredWorkflowStep::FinalizeWorkflow { reason } => {
-                finalize_workflow_completion(manager, session_id, reason).await?;
+                finalize_workflow_completion(session_repo, active_sessions, app_handle, session_id, reason).await?;
             }
         },
         CompactionResumeAction::ResumeCompletion => {
@@ -224,10 +251,79 @@ pub async fn handle_compact_response(
                 "▶️ Resuming blocked LLM completion after compaction for session {}",
                 session_id
             );
-            spawn_resume_completion(manager, session_id, "LLM completion");
+            spawn_resume_completion(session_repo, active_sessions, proxy_manager, app_handle, session_id, "LLM completion");
         }
         CompactionResumeAction::Nothing => {}
     }
 
     Ok(())
+}
+
+pub async fn get_compact_context(
+    active_sessions: &Arc<RwLock<HashMap<String, AgentSession>>>,
+    session_id: &str,
+) -> Result<Option<CompactContextRecord>, String> {
+    let active = active_sessions.read().await;
+    if let Some(session) = active.get(session_id) {
+        let compact = session.compact_context.read().await;
+        if compact.is_some() {
+            return Ok((*compact).clone());
+        }
+    }
+
+    let repo = crate::state::get_compact_context_repository();
+    repo.get_by_session_id(session_id)
+        .await
+        .map_err(|e| e.to_string())
+}
+
+pub async fn save_compact_context(
+    active_sessions: &Arc<RwLock<HashMap<String, AgentSession>>>,
+    session_id: &str,
+    record: CompactContextRecord,
+) -> Result<(), String> {
+    {
+        let active = active_sessions.read().await;
+        if let Some(session) = active.get(session_id) {
+            let mut compact = session.compact_context.write().await;
+            *compact = Some(record.clone());
+        }
+    }
+
+    let repo = crate::state::get_compact_context_repository();
+    repo.upsert(&record).await.map_err(|e| e.to_string())?;
+
+    Ok(())
+}
+
+pub async fn wait_for_compaction_to_settle(
+    active_sessions: &Arc<RwLock<HashMap<String, AgentSession>>>,
+    session_id: &str,
+    timeout: std::time::Duration,
+) -> Result<(), String> {
+    let started_at = std::time::Instant::now();
+
+    loop {
+        let is_settled = {
+            let active = active_sessions.read().await;
+            let session = active
+                .get(session_id)
+                .ok_or_else(|| format!("Session not found: {}", session_id))?;
+            session.compaction.is_settled()
+        };
+
+        if is_settled {
+            return Ok(());
+        }
+
+        if started_at.elapsed() >= timeout {
+            return Err(format!(
+                "Timed out waiting for compaction to settle for session {} after {} seconds",
+                session_id,
+                timeout.as_secs()
+            ));
+        }
+
+        tokio::time::sleep(std::time::Duration::from_millis(150)).await;
+    }
 }
